+++ internals/scripts/CheckBuildsExist.js
@@ -3,32 +3,28 @@ import path from 'path';
 import chalk from 'chalk';
 import fs from 'fs';
 
-function CheckBuildsExist() {
-  const mainPath = path.join(__dirname, '..', '..', 'app', 'main.prod.js');
-  const rendererPath = path.join(
-    __dirname,
-    '..',
-    '..',
-    'app',
-    'dist',
-    'renderer.prod.js'
+const mainPath = path.join(__dirname, '..', '..', 'app', 'main.prod.js');
+const rendererPath = path.join(
+  __dirname,
+  '..',
+  '..',
+  'app',
+  'dist',
+  'renderer.prod.js'
+);
+
+if (!fs.existsSync(mainPath)) {
+  throw new Error(
+    chalk.whiteBright.bgRed.bold(
+      'The main process is not built yet. Build it by running "yarn build-main"'
+    )
   );
-
-  if (!fs.existsSync(mainPath)) {
-    throw new Error(
-      chalk.whiteBright.bgRed.bold(
-        'The main process is not built yet. Build it by running "yarn build-main"'
-      )
-    );
-  }
-
-  if (!fs.existsSync(rendererPath)) {
-    throw new Error(
-      chalk.whiteBright.bgRed.bold(
-        'The renderer process is not built yet. Build it by running "yarn build-renderer"'
-      )
-    );
-  }
 }
 
-CheckBuildsExist();
+if (!fs.existsSync(rendererPath)) {
+  throw new Error(
+    chalk.whiteBright.bgRed.bold(
+      'The renderer process is not built yet. Build it by running "yarn build-renderer"'
+    )
+  );
+}